Arrange CI

+++ integrate/runyig.sh
@@ -1,6 +1,8 @@
 BASEDIR=$(dirname $(pwd))
-sudo docker run --rm -d --name yig \
-			-p 80:80 \
+echo ${BASEDIR}
+if [ -x "$BASEDIR/yig" ]; then 
+    sudo docker run --rm -d --name yig \
+			 -p 80:80 \
 	                 -p 9000:9000 \
 			 -v ${BASEDIR}/integrate/cephconf:/etc/ceph/ \
 			 -v ${BASEDIR}/integrate/yigconf:/etc/yig/ \
@@ -8,4 +10,17 @@ sudo docker run --rm -d --name yig \
 			 -v ${BASEDIR}:/work \
                          --net=integrate_vpcbr \
                          --ip 10.5.0.18 \
-			 yig /work/build/bin/yig
+			 journeymidnight/yig /work/yig
+    echo "started yig from local dir"
+else
+    sudo docker run --rm -d --name yig \
+			 -p 80:80 \
+	                 -p 9000:9000 \
+			 -v ${BASEDIR}/integrate/cephconf:/etc/ceph/ \
+			 -v ${BASEDIR}/integrate/yigconf:/etc/yig/ \
+			 -v ${BASEDIR}:/var/log/yig \
+                         --net=integrate_vpcbr \
+                         --ip 10.5.0.18 \
+			 journeymidnight/yig
+    echo "started yig from docker.io/journeymidnight/yig"
+fi
